// frontend/generated-api/models/tribe.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Survivor Fantasy League - OpenAPI 3.0.3
 * This is the Survivor Fantasy League Spec for API's
 *
 * OpenAPI spec version: 1.0.11
 * Contact: pklein111697@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Color } from './color';
import { Episode } from './episode';
 /**
 * 
 *
 * @export
 * @interface Tribe
 */
export interface Tribe {

    /**
     * The season's unique identifier
     *
     * @type {string}
     * @memberof Tribe
     * @example 47
     */
    seasonId: string;

    /**
     * The tribe's unique identifier
     *
     * @type {string}
     * @memberof Tribe
     * @example 1
     */
    id: string;

    /**
     * The tribe's name
     *
     * @type {string}
     * @memberof Tribe
     * @example Luzon
     */
    name: string;

    /**
     * @type {Color}
     * @memberof Tribe
     */
    color: Color;

    /**
     * The tribe's image URL
     *
     * @type {string}
     * @memberof Tribe
     * @example https://www.cbs.com/shows/survivor/photos/1007124/season-28-cast-photos/1007124_1_8x6.jpg
     */
    imageUrl: string;

    /**
     * Whether the tribe is a merge tribe
     *
     * @type {boolean}
     * @memberof Tribe
     * @example false
     */
    isMergeTribe?: boolean;

    /**
     * @type {Episode}
     * @memberof Tribe
     */
    episodeStarted?: Episode;
}